WIP: let min/max use block reduce but tests are failing, move tests according to PR discussion
